Add files via upload
Added the account_DB sql file, which includes a table titled user, with columns email(key),password, first_name, last_name, is_seller. Note that we are missing a way to link the users listed products, but Beau is working on that. Beau thinks the best(or at least easiest) way to do that will be to create a new column in the user table that's a string of product ID's posted by that user, and that string get's handled by some JavaScript. The issue is that you can't create a column of array's in SQL. Will update DB with this idea unless a better one is proposed.